# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TASK MANAGER API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/add/course": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "3d4e453f1c0df4d6b5b317af1641238d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "BackEnd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/update/course/{id}": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "a8a683a1fbbcef71ab8606a29f89321c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste task id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "BackEnd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete/course/{id}": {
            "delete": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "e052aa1901ee6ccef56f495ba9bbcfbb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste task id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "BackEnd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}